requirements_core.txt
# ============================================
# CORE REQUIREMENTS - Fraud Detection API
# ============================================

# Flask Framework
Flask>=3.0.0
Flask-CORS>=4.0.0

# OpenAI (for OCR parsing)
openai>=1.3.0

# OCR Libraries
pytesseract>=0.3.10
Pillow>=10.2.0

# Data Processing - FIXED VERSIONS
numpy==1.26.4
pandas>=2.1.0
scipy==1.13.1

# Machine Learning - FIXED VERSIONS
scikit-learn==1.2.2
xgboost==2.0.3
imbalanced-learn==0.10.1
joblib>=1.3.0
threadpoolctl==2.1.0

# Plotting (optional - for visualization)
matplotlib>=3.8.0

# Environment Variables
python-dotenv>=1.0.0

# ============================================
# OPTIONAL - Development & Testing
# ============================================

# API Documentation
flask-swagger-ui>=4.11.1

# Testing
pytest>=7.4.0
pytest-flask>=1.2.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
